Add support for scraping and handling events in addition to movies
The Scraper class has been abstracted and enhanced to handle not only movies but also events. Additionally, the langdetect library was imported to recognize language of the events. Few workflow variables were updated. The MovieScraper and a newly introduced EventsScraper classes inherit from the abstract Scraper class. This refactoring is done to increase the code's versatility and improve its ability to handle different types of sources (movies and events).